% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{HPT_feedback_with_graph}
\alias{HPT_feedback_with_graph}
\title{HPT feedback (with graph)}
\usage{
HPT_feedback_with_graph(dict = HPT::HPT_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textual and graphical feedback at the end of the test.
}
\examples{
\dontrun{
HPT_demo(feedback = HPT_feedback_with_score())}
}
